# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{organizationId}/tickets/{ticketId}/attachments/init": {
      "post": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "AttachmentsController_initiate",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateAttachmentUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiatedAttachment"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Attachments: initiate",
        "tags": [
          "Attachments"
        ]
      }
    },
    "/v1/organizations/{organizationId}/tickets/{ticketId}/attachments/{attachmentId}/complete": {
      "post": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "AttachmentsController_complete",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Attachments: complete",
        "tags": [
          "Attachments"
        ]
      }
    },
    "/v1/organizations/{organizationId}/tickets/{ticketId}/attachments/{attachmentId}/download": {
      "get": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:read. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "AttachmentsController_download",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Download"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Attachments: download",
        "tags": [
          "Attachments"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpsDesk API",
    "description": "Tenant-scoped support operations API. Use the Next.js BFF for browser requests; Auth0 bearer tokens belong on the API boundary. See docs/api/README.md for workflows and Socket.IO events. No real credentials or account details are embedded.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "Local API only; configure deployment URLs privately"
    }
  ],
  "components": {
    "securitySchemes": {
      "auth0": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Auth0 access token with the configured API audience and issuer; never an ID token."
      },
      "svix": {
        "type": "apiKey",
        "in": "header",
        "name": "svix-signature",
        "description": "Verified Resend/Svix signature. Also requires svix-id and svix-timestamp; the signature covers the exact raw body."
      }
    },
    "schemas": {
      "InitiateAttachmentUploadDto": {
        "type": "object",
        "properties": {
          "originalName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "contentType": {
            "type": "string",
            "enum": [
              "application/pdf",
              "image/jpeg",
              "image/png",
              "image/webp",
              "image/gif",
              "text/plain",
              "text/csv",
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
              "application/vnd.openxmlformats-officedocument.presentationml.presentation"
            ]
          },
          "sizeBytes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 26214400
          }
        },
        "required": [
          "originalName",
          "contentType",
          "sizeBytes"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "originalName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "UPLOADED"
            ]
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "originalName",
          "contentType",
          "sizeBytes",
          "status",
          "uploadedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "InitiatedAttachment": {
        "type": "object",
        "properties": {
          "attachment": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "originalName": {
                "type": "string"
              },
              "contentType": {
                "type": "string"
              },
              "sizeBytes": {
                "type": "integer"
              },
              "status": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "UPLOADED"
                ]
              },
              "uploadedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "originalName",
              "contentType",
              "sizeBytes",
              "status",
              "uploadedAt",
              "createdAt",
              "updatedAt"
            ]
          },
          "upload": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri",
                "description": "Short-lived signed URL; treat as a secret."
              },
              "method": {
                "type": "string",
                "enum": [
                  "PUT"
                ]
              },
              "headers": {
                "type": "object",
                "properties": {
                  "Content-Type": {
                    "type": "string"
                  },
                  "x-amz-meta-attachment-id": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "Content-Type",
                  "x-amz-meta-attachment-id"
                ]
              },
              "expiresInSeconds": {
                "type": "integer"
              }
            },
            "required": [
              "url",
              "method",
              "headers",
              "expiresInSeconds"
            ]
          }
        },
        "required": [
          "attachment",
          "upload"
        ]
      },
      "Download": {
        "type": "object",
        "properties": {
          "attachment": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "originalName": {
                "type": "string"
              },
              "contentType": {
                "type": "string"
              },
              "sizeBytes": {
                "type": "integer"
              },
              "status": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "UPLOADED"
                ]
              },
              "uploadedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "messageId": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "id",
              "originalName",
              "contentType",
              "sizeBytes",
              "status",
              "uploadedAt",
              "messageId"
            ]
          },
          "download": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              },
              "expiresInSeconds": {
                "type": "integer"
              }
            },
            "required": [
              "url",
              "expiresInSeconds"
            ]
          }
        },
        "required": [
          "attachment",
          "download"
        ]
      },
...
    }
  }
}